auth logic sent to new component
